MB-67520: Fix narrowing warnings in durability code
Make changes to address narrowing conversions mostly due to the
choice of uint8_t for counting replica chains (how many nodes
a vbucket can be replica on).

The use of uint8_t internally seems like a premature
space-optimisation. To keep the change simple, the uint8_t typing is
retained, but various narrow casts and tweaks should provide sane
behaviour (failing to intialise and lose precision and failing before
overflow).

Validation of replication chain (JSON value from setVbState) limits
them to maxReplicas+1 making the use of u8 for chain counting
appearing to be safe.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230761
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/durability/active_durability_monitor.cc
@@ -668,12 +668,14 @@ size_t ActiveDurabilityMonitor::getTotalMemoryUsed() const {
 
 uint8_t ActiveDurabilityMonitor::getFirstChainSize() const {
     const auto s = state.rlock();
-    return s->firstChain ? s->firstChain->positions.size() : 0;
+    return gsl::narrow_cast<uint8_t>(
+            s->firstChain ? s->firstChain->positions.size() : 0);
 }
 
 uint8_t ActiveDurabilityMonitor::getSecondChainSize() const {
     const auto s = state.rlock();
-    return s->secondChain ? s->secondChain->positions.size() : 0;
+    return gsl::narrow_cast<uint8_t>(
+            s->secondChain ? s->secondChain->positions.size() : 0);
 }
 
 uint8_t ActiveDurabilityMonitor::getFirstChainMajority() const {
